feat: Rotation Suffix Time Format
* Rotation Suffix Time Format

Added RotationFileSuffixTimeFormat field to Logger

default: 2006-01-02T15-04-05.000

It used for naming the rotated log files. If not set, then default value is used.

Added functions to count the precision of the layout passed and test the validity of the supplied layout.
If the layout supplied is not valid then default is used

* Validation Flag and comment fix.

added flag to prevent multiple validation checks

* feat(rotation): Custom Time Format for a rotated file

Added RotationFileSuffixTimeFormat field to Logger

default: 2006-01-02T15-04-05.000

Using this feature, you can now set up your own time-format that'll be
used as suffix of a rotated file.

* feat!(rotation file name): BackupTimeFormat is now required

BREAKING CHANGE: `BackupTimeFormat` field is now required for Logger
instance to work. Returns error if invalid value is passed.

* feat(rotate): Optional field for backup timeformat is added.

When rotating, if a valid `BackupTimeFormat` is configured then it is
used.
Otherwise, we silently fall back to using default timeformat : `2006-01-02T15-04-05.000`

* fix: removed unused functions

* fix: Code style consistency changes

1. Added log to Stderr after fallback to default time format
2. `timeFromName` function uses supplied format and fallbacks to default
   if no provided

---------

Co-authored-by: Dean Ruina <81315494+DeRuina@users.noreply.github.com>